Take time signature from the midi
Significant change.  The configuration options remove time signature and replace it with midi track number.  The time signature is then worked out heuristically from the track.  So the startup parameters now differ.  Added the first sample that was created from a midi keyboard through a sequencer.  This produces abc which is currently unusable because the program is fooled into producing spurious triplets because the notes are not of an accurate duration.

Ready now for phase II