Added note articulation and some note filtering
This commit starts to take account of note durations that are too small to measure or where notes overlap.  The articulate function produces a more legato-like score where notes can smoothely elide but not overlap.  This should simplify score production for human-produced performances.

Improved the handling of ties in bar lines so that if a note needs to be tied across bars, this only takes place if the notes on both sides of the bar are long enough to display and notate.  Otherwise the note is truncated.

Added a generic filter over scores.