Improved the articulate function
This uses heuristics to try to simplify the melody.  Any note that is overlapped by its predeccessor is curtailed or removed entirely (which has the side effect of removing chords). Any note which falls short of its successor by an amount less than the smallest detectable note duration is extended.  The idea is to give a more legato-like feel.