Keep track of the beat number
Up till now, we only keep track of whether or not a note lies on a beat.  Now we keep track of both this fact and also its beat number when it is on a beat using Maybe BeatNo.  This allows us to beam together notes more accurately in the score (particularly needed for tunes in 4/4)