Add recognition of rests
This is tricky because rests are not represented in MIDI other than by an absence - a gap between adjacent notes.  This could be caused by a note decaying more quickly than intended or could be a legitimate rest.  The articulate function takes the view that if such a gap is less than the smallest detectable note, the previous note is extended, otherwise it is notated as a rest.  The split long notes function must now detect and split rests in pretty much the same manner as it splits notes.